backend/requirements.txt
fastapi
uvicorn[standard]
httpx
dnspython
sqlalchemy
alembic
python-dotenv
websockets